+++ internal/pkg/node/datastructure.go
@@ -3,6 +3,7 @@ package node
 import (
 	"net"
 
+	"github.com/warewulf/warewulf/internal/pkg/util"
 	"github.com/warewulf/warewulf/internal/pkg/wwtype"
 )
 
@@ -93,7 +94,7 @@ Holds the disks of a node
 */
 type Disk struct {
 	id         string                `yaml:"-"                    json:"-"`
-	WipeTable  wwtype.WWbool         `yaml:"wipe_table,omitempty" json:"wipe_table,omitempty" lopt:"diskwipe" comment:"whether or not the partition tables shall be wiped"`
+	WipeTableP *bool                 `yaml:"wipe_table,omitempty" json:"wipe_table,omitempty" lopt:"diskwipe" comment:"whether or not the partition tables shall be wiped"`
 	Partitions map[string]*Partition `yaml:"partitions,omitempty" json:"partitions,omitempty"`
 }
 
@@ -102,29 +103,52 @@ partition definition, the label must be uniq so its used as the key in the
 Partitions map
 */
 type Partition struct {
-	id                 string        `yaml:"-"                              json:"-"`
-	Number             string        `yaml:"number,omitempty"               json:"number,omitempty"               lopt:"partnumber" comment:"Set the partition number, if not set next free slot is used" type:"uint"`
-	SizeMiB            string        `yaml:"size_mib,omitempty"             json:"size_mib,omitempty"             lopt:"partsize"   comment:"Set the size of the partition, if not set maximal possible size is used" type:"uint"`
-	StartMiB           string        `yaml:"start_mib,omitempty"            json:"start_mib,omitempty"                              comment:"the start of the partition" type:"uint"`
-	TypeGuid           string        `yaml:"type_guid,omitempty"            json:"type_guid,omitempty"            lopt:"parttype"   comment:"Set the partition type GUID"`
-	Guid               string        `yaml:"guid,omitempty"                 json:"guid,omitempty"                                   comment:"the GPT unique partition GUID"`
-	WipePartitionEntry wwtype.WWbool `yaml:"wipe_partition_entry,omitempty" json:"wipe_partition_entry,omitempty"                   comment:"if true, Ignition will clobber an existing partition if it does not match the config"`
-	ShouldExist        wwtype.WWbool `yaml:"should_exist,omitempty"         json:"should_exist,omitempty"         lopt:"partcreate" comment:"Create partition if it does not exist"`
-	Resize             wwtype.WWbool `yaml:"resize,omitempty"               json:"resize,omitempty"                                 comment:"whether or not the existing partition should be resize"`
+	id                  string `yaml:"-"                              json:"-"`
+	Number              string `yaml:"number,omitempty"               json:"number,omitempty"               lopt:"partnumber" comment:"Set the partition number, if not set next free slot is used" type:"uint"`
+	SizeMiB             string `yaml:"size_mib,omitempty"             json:"size_mib,omitempty"             lopt:"partsize"   comment:"Set the size of the partition, if not set maximal possible size is used" type:"uint"`
+	StartMiB            string `yaml:"start_mib,omitempty"            json:"start_mib,omitempty"                              comment:"the start of the partition" type:"uint"`
+	TypeGuid            string `yaml:"type_guid,omitempty"            json:"type_guid,omitempty"            lopt:"parttype"   comment:"Set the partition type GUID"`
+	Guid                string `yaml:"guid,omitempty"                 json:"guid,omitempty"                                   comment:"the GPT unique partition GUID"`
+	WipePartitionEntryP *bool  `yaml:"wipe_partition_entry,omitempty" json:"wipe_partition_entry,omitempty"                   comment:"if true, Ignition will clobber an existing partition if it does not match the config"`
+	ShouldExistP        *bool  `yaml:"should_exist,omitempty"         json:"should_exist,omitempty"         lopt:"partcreate" comment:"Create partition if it does not exist"`
+	ResizeP             *bool  `yaml:"resize,omitempty"               json:"resize,omitempty"                                 comment:"whether or not the existing partition should be resize"`
 }
 
 /*
 Definition of a filesystem. The device is uniq so its used as key
 */
 type FileSystem struct {
-	id             string        `yaml:"-"                         json:"-"`
-	Format         string        `yaml:"format,omitempty"          json:"format,omitempty"          lopt:"fsformat" comment:"format of the file system"`
-	Path           string        `yaml:"path,omitempty"            json:"path,omitempty"            lopt:"fspath"   comment:"the mount point of the file system"`
-	WipeFileSystem wwtype.WWbool `yaml:"wipe_filesystem,omitempty" json:"wipe_filesystem,omitempty" lopt:"fswipe"   comment:"wipe file system at boot"`
-	Label          string        `yaml:"label,omitempty"           json:"label,omitempty"                           comment:"the label of the filesystem"`
-	Uuid           string        `yaml:"uuid,omitempty"            json:"uuid,omitempty"                            comment:"the uuid of the filesystem"`
-	Options        []string      `yaml:"options,omitempty"         json:"options,omitempty"                         comment:"any additional options to be passed to the format-specific mkfs utility"`
-	MountOptions   string        `yaml:"mount_options,omitempty"   json:"mount_options,omitempty"                   comment:"any special options to be passed to the mount command"`
+	id              string   `yaml:"-"                         json:"-"`
+	Format          string   `yaml:"format,omitempty"          json:"format,omitempty"          lopt:"fsformat" comment:"format of the file system"`
+	Path            string   `yaml:"path,omitempty"            json:"path,omitempty"            lopt:"fspath"   comment:"the mount point of the file system"`
+	WipeFileSystemP *bool    `yaml:"wipe_filesystem,omitempty" json:"wipe_filesystem,omitempty" lopt:"fswipe"   comment:"wipe file system at boot"`
+	Label           string   `yaml:"label,omitempty"           json:"label,omitempty"                           comment:"the label of the filesystem"`
+	Uuid            string   `yaml:"uuid,omitempty"            json:"uuid,omitempty"                            comment:"the uuid of the filesystem"`
+	Options         []string `yaml:"options,omitempty"         json:"options,omitempty"                         comment:"any additional options to be passed to the format-specific mkfs utility"`
+	MountOptions    string   `yaml:"mount_options,omitempty"   json:"mount_options,omitempty"                   comment:"any special options to be passed to the mount command"`
 }
 
 type Resource interface{}
+
+// Disk methods
+func (disk *Disk) WipeTable() bool {
+	return util.BoolP(disk.WipeTableP)
+}
+
+// Partition methods
+func (partition *Partition) WipePartitionEntry() bool {
+	return util.BoolP(partition.WipePartitionEntryP)
+}
+
+func (partition *Partition) ShouldExist() bool {
+	return util.BoolP(partition.ShouldExistP)
+}
+
+func (partition *Partition) Resize() bool {
+	return util.BoolP(partition.ResizeP)
+}
+
+// FileSystem methods
+func (fs *FileSystem) WipeFileSystem() bool {
+	return util.BoolP(fs.WipeFileSystemP)
+}